requirements.txt
requests  
pillow  
